refactor: notes to PR#3
