+++ spring-03/src/main/java/ru/korolkovrs/spring/i18n_util/ResourcePathResolver.java
@@ -13,6 +13,7 @@ import java.util.Map;
 public class ResourcePathResolver {
     private final LocaleContext localeContext;
 
+
     private Map<String, String> codes = new HashMap<>();
 
     public ResourcePathResolver(LocaleContext localeContext) {
