+++ spring-03/src/main/java/ru/korolkovrs/spring/service/IOInternatiolizeService.java
@@ -1,0 +1,7 @@
+package ru.korolkovrs.spring.service;
+
+public interface IOInternatiolizeService {
+    String input();
+    void out(String s, Object... args);
+    void outWithInternalize(String s, Object... args);
+}
