+++ spring-03/src/main/java/ru/korolkovrs/spring/service/IOInternatiolizeServiceImpl.java
@@ -1,24 +1,26 @@
-package ru.korolkovrs.spring.policy;
+package ru.korolkovrs.spring.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import ru.korolkovrs.spring.i18n_util.Internalizer;
-import ru.korolkovrs.spring.service.IOServiceImpl;
 
-@Component
+@Service
 @RequiredArgsConstructor
-public class IOServiceInternatiolizePolicy {
-    private final IOServiceImpl ioService;
+public class IOInternatiolizeServiceImpl implements IOInternatiolizeService {
+    private final IOService ioService;
     private final Internalizer internalizer;
 
+    @Override
     public String input() {
         return ioService.input();
     }
 
+    @Override
     public void out(String s, Object... args) {
         ioService.out(s);
     }
 
+    @Override
     public void outWithInternalize(String s, Object... args) {
         ioService.out(internalizer.internalizeMessage(s), args);
     }
