+++ spring-03/src/main/java/ru/korolkovrs/spring/service/TestServiceImpl.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Service;
 import ru.korolkovrs.spring.converter.QuestionToStringConverter;
 import ru.korolkovrs.spring.domain.Question;
 import ru.korolkovrs.spring.domain.User;
-import ru.korolkovrs.spring.policy.IOServiceInternatiolizePolicy;
 
 import java.util.List;
 
@@ -15,7 +14,7 @@ public class TestServiceImpl implements TestService {
     private final UserService userService;
     private final QuestionService questionService;
     private final QuestionToStringConverter converter;
-    private final IOServiceInternatiolizePolicy ioServiceInternatiolizePolicy;
+    private final IOInternatiolizeServiceImpl ioInternatiolizeServiceImpl;
 
 
     @Override
@@ -25,13 +24,13 @@ public class TestServiceImpl implements TestService {
         int score = 0;
 
         for (Question question : questions) {
-            ioServiceInternatiolizePolicy.out(converter.convert(question));
+            ioInternatiolizeServiceImpl.out(converter.convert(question));
             int answer = waitCorrectInputAndGet(question);
             if (question.getAnswers().get(answer).isCorrect()) {
                 score++;
             }
         }
-        ioServiceInternatiolizePolicy.outWithInternalize("testService.test_result",
+        ioInternatiolizeServiceImpl.outWithInternalize("testService.test_result",
                 user.getName(), user.getSurname(), score);
     }
 
@@ -45,14 +44,14 @@ public class TestServiceImpl implements TestService {
 
     private int checkUserAnswer(Question q) {
         try {
-            ioServiceInternatiolizePolicy.outWithInternalize("testService.enter_answer", q.getAnswers().size());
-            int answer = Integer.parseInt(ioServiceInternatiolizePolicy.input());
+            ioInternatiolizeServiceImpl.outWithInternalize("testService.enter_answer", q.getAnswers().size());
+            int answer = Integer.parseInt(ioInternatiolizeServiceImpl.input());
             if (answer > 0 && answer <= q.getAnswers().size()) {
                 return answer - 1;
             }
             throw new RuntimeException();
         } catch (RuntimeException e) {
-            ioServiceInternatiolizePolicy.outWithInternalize("testService.incorrect_format");
+            ioInternatiolizeServiceImpl.outWithInternalize("testService.incorrect_format");
             return -1;
         }
     }
